Start work on hints and reveal answer
